+++ processing/threedidepth_algorithms.py
@@ -20,6 +20,7 @@ from qgis.core import QgsProcessingException
 from qgis.core import QgsProcessingParameterBoolean
 from qgis.core import QgsProcessingParameterEnum
 from qgis.core import QgsProcessingParameterFile
+from qgis.core import QgsProcessingParameterFileDestination
 from qgis.core import QgsProcessingParameterNumber
 from qgis.core import QgsProcessingParameterRasterLayer
 from qgis.core import QgsProcessingParameterString
@@ -39,6 +40,7 @@ from threedi_results_analysis.utils.user_messages import pop_up_info
 import h5py
 import logging
 import os
+import pathlib
 
 
 logger = logging.getLogger(__name__)
@@ -357,6 +359,126 @@ class ThreediDepthAlgorithm(QgsProcessingAlgorithm):
             layer = QgsRasterLayer(waterdepth_output_file, raster_filename)
         context.temporaryLayerStore().addMapLayer(layer)
         layer_details = QgsProcessingContext.LayerDetails(raster_filename, context.project(), self.WATER_DEPTH_OUTPUT)
+        context.addLayerToLoadOnCompletion(layer.id(), layer_details)
+        return {}
+
+
+class ThreediMaxDepthAlgorithm(QgsProcessingAlgorithm):
+    """
+    Calculates maximum water depth or water level rasters from 3Di result NetCDF
+    """
+
+    # Constants used to refer to parameters and outputs. They will be
+    # used when calling the algorithm from another algorithm, or when
+    # calling from the QGIS console.
+
+    MODES = [
+        Mode(MODE_LIZARD, "Interpolated water depth"),
+        Mode(MODE_LIZARD_S1, "Interpolated water level"),
+        Mode(MODE_CONSTANT, "Non-interpolated water depth"),
+        Mode(MODE_CONSTANT_S1, "Non-interpolated water level"),
+    ]
+
+    GRIDADMIN_INPUT = "GRIDADMIN_INPUT"
+    RESULTS_3DI_INPUT = "RESULTS_3DI_INPUT"
+    DEM_INPUT = "DEM_INPUT"
+    MODE_INPUT = "MODE_INPUT"
+    OUTPUT_FILENAME = "OUTPUT_FILENAME"
+    WATER_DEPTH_OUTPUT = "WATER_DEPTH_OUTPUT"
+
+    def tr(self, string):
+        """
+        Returns a translatable string with the self.tr() function.
+        """
+        return QCoreApplication.translate("Processing", string)
+
+    def createInstance(self):
+        return ThreediMaxDepthAlgorithm()
+
+    def name(self):
+        """Returns the algorithm name, used for identifying the algorithm"""
+        return "threedidepth_max"
+
+    def displayName(self):
+        """
+        Returns the translated algorithm name, which should be used for any
+        user-visible display of the algorithm name.
+        """
+        return self.tr("Maximum water depth/level raster")
+
+    def group(self):
+        """Returns the name of the group this algorithm belongs to"""
+        return self.tr("Post-process results")
+
+    def groupId(self):
+        """Returns the unique ID of the group this algorithm belongs to"""
+        return "postprocessing"
+
+    def shortHelpString(self):
+        """Returns a localised short helper string for the algorithm"""
+        return self.tr("Calculate maximum water depth/level raster over all timesteps")
+
+    def initAlgorithm(self, config=None):
+        """Here we define the inputs and output of the algorithm"""
+        # Input parameters
+        self.addParameter(
+            QgsProcessingParameterFile(self.GRIDADMIN_INPUT, self.tr("Gridadmin.h5 file"), extension="h5")
+        )
+        self.addParameter(
+            QgsProcessingParameterFile(self.RESULTS_3DI_INPUT, self.tr("Results_3di.nc file"), extension="nc")
+        )
+        self.addParameter(QgsProcessingParameterRasterLayer(self.DEM_INPUT, self.tr("DEM")))
+        self.addParameter(
+            QgsProcessingParameterEnum(
+                name=self.MODE_INPUT,
+                description=self.tr("Interpolation mode"),
+                options=[m.description for m in self.MODES],
+                defaultValue=MODE_LINEAR,
+            )
+        )
+        self.addParameter(
+            QgsProcessingParameterFileDestination(
+                self.OUTPUT_FILENAME,
+                self.tr("Destination file path for water depth/level raster"),
+                fileFilter="*.tif",
+            )
+        )
+
+    def processAlgorithm(self, parameters, context, feedback):
+        """
+        Create the water depth raster with the provided user inputs
+        """
+        dem_filename = self.parameterAsRasterLayer(parameters, self.DEM_INPUT, context).source()
+        gridadmin_path = parameters[self.GRIDADMIN_INPUT]
+        results_3di_path = parameters[self.RESULTS_3DI_INPUT]
+        mode_index = self.parameterAsEnum(parameters, self.MODE_INPUT, context)
+
+        waterdepth_output_file = parameters[self.OUTPUT_FILENAME]
+        if not waterdepth_output_file:
+            raise QgsProcessingException(self.invalidSourceError(parameters, self.OUTPUT_FILENAME))
+
+        layer_name = pathlib.Path(waterdepth_output_file).stem
+
+        if os.path.isfile(waterdepth_output_file):
+            os.remove(waterdepth_output_file)
+
+        try:
+            calculate_waterdepth(
+                gridadmin_path=gridadmin_path,
+                results_3di_path=results_3di_path,
+                dem_path=dem_filename,
+                waterdepth_path=waterdepth_output_file,
+                calculate_maximum_waterlevel=True,
+                mode=self.MODES[mode_index].name,
+                progress_func=Progress(feedback),
+            )
+        except CancelError:
+            # When the process is cancelled, we just show the intermediate product
+            pass
+
+        layer = QgsRasterLayer(waterdepth_output_file, layer_name)
+        context.temporaryLayerStore().addMapLayer(layer)
+        layer_details = QgsProcessingContext.LayerDetails(layer_name, context.project(), self.WATER_DEPTH_OUTPUT)
         context.addLayerToLoadOnCompletion(layer.id(), layer_details)
         return {}