+++ CHANGES.rst
@@ -34,6 +34,10 @@
 - Result Manager: Added option for users to right-click on any item in the list and delete it via the context menu (#844)
 - Removed Cache Clearer tool and Result Selection tool (#843)
 
+- Bump threedidepth to 0.6.1
+- Add algorithm for maximum waterdepth to processing toolbox.
+- Make the plugin work with all QGis versions by making installed
+  h5py version depend on QGis version.
 
 2.5.3 (2023-06-16)
 ------------------